// core/sj-frontend/src/client/app/modules/module.ts
export class Module {
    'module-type': string;
    'module-name': string;
    'module-version': string;
    name: string;
    description: string;
    version: string;
    author: string;
    license: string;
    'engine-name': string;
    'engine-version': string;
    'validator-class': string;
    'executor-class': string;
}